requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
astroid==3.3.6
attrs==24.3.0
colorama==0.4.6
dill==0.3.9
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.2
pytest==8.3.4
tomlkit==0.13.2
